Split auth config for Edge runtime, downgrade Prisma to 6.x for Mongo support
middleware.ts runs on Next's Edge runtime and was importing the full
auth() config (PrismaAdapter, Credentials.authorize), which pulled
Prisma's DB-querying code into the Edge bundle and broke yarn dev.
Split into auth.config.ts (Edge-safe: Google provider + token->session
mapping) and auth.ts (full config with PrismaAdapter and Credentials
provider); middleware.ts now builds its own NextAuth instance from the
Edge-safe config.

Prisma 7.8.0 has no MongoDB driver adapter, so it could never connect;
downgraded prisma/@prisma/client to ^6.19.3 and restored the seed
script entry that got affected during testing.

Also stop tracking .env (contains live DB/API credentials) and add it
to .gitignore.
