--- .env
@@ -1,25 +1,0 @@
-AWS_BUCKET=wedin-images-dev
-AWS_BUCKET_REGION=us-east-1
-AWS_ACCESS_KEY_ID=AKIATCKAPQZJJTWFXKFK
-AWS_SECRET_ACCESS_KEY=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
-
-GOOGLE_CLIENT_ID=351097060218-rm5i0drr64g6dcb5d2vuammdh9iq9vjr.apps.googleusercontent.com
-GOOGLE_CLIENT_SECRET=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
-
-FACEBOOK_CLIENT_ID=925826102595379
-FACEBOOK_CLIENT_SECRET=728dfd4ce30bcf4dd3966f146d97ab57
-
-EMAIL_SERVER_HOST=smtp.resend.com
-EMAIL_SERVER_PORT=465
-EMAIL_SERVER_USER=resend
-EMAIL_SERVER_PASSWORD=re_GpiNfd8r_6UXo8F761fBiB1dE5KQ4CgwX
-EMAIL_FROM=wedin@no-reply.com
-
-DATABASE_URL="mongodb+srv://dev:26xFmFK4Jfd7B1aJ@clusterwedin.hpiycyj.mongodb.net/clusterwedin?retryWrites=true&w=majority"
-NODE_ENV=development
-
-NEXTAUTH_URL=http://localhost:3000/
-NEXTAUTH_SECRET=s03CqB/bJxPq01jgdNijuVMGSRov7ihiNqiy/9w35H8=
-
-RESEND_API_KEY=re_TXXGq52g_3gpX9sFgJpZEh6pS5TtdDztm
-AUTH_RESEND_KEY=re_aP2ZLeUp_QHgPYMuV4ToUmPDSCGoYy66U
