+++ auth.config.ts
@@ -1,57 +1,36 @@
-import prismaClient from '@/prisma/client';
-import bcrypt from 'bcryptjs';
-import Credentials from 'next-auth/providers/credentials';
-// import Facebook from 'next-auth/providers/facebook';
 import Google from 'next-auth/providers/google';
 import type { NextAuthConfig } from 'next-auth';
-import { PrismaAdapter } from '@auth/prisma-adapter';
-import { LoginSchema } from '@/schemas/auth';
 
-const authOptions: NextAuthConfig = {
-  adapter: PrismaAdapter(prismaClient),
+// Edge-safe config: no Prisma/adapter/DB calls here.
+// This is used directly by middleware.ts (Edge runtime) and merged
+// into the full config in auth.ts (Node.js runtime).
+const authConfig: NextAuthConfig = {
   providers: [
     Google({
       clientId: process.env.GOOGLE_CLIENT_ID as string,
       clientSecret: process.env.GOOGLE_CLIENT_SECRET as string,
       allowDangerousEmailAccountLinking: true,
     }),
-    // Facebook({
-    //   clientId: process.env.GOOGLE_CLIENT_ID as string,
-    //   clientSecret: process.env.GOOGLE_CLIENT_SECRET as string,
-    //   allowDangerousEmailAccountLinking: true,
-    // }),
-    Credentials({
-      async authorize(credentials) {
-        const validatedFields = LoginSchema.safeParse(credentials);
-
-        if (validatedFields.success) {
-          const { email, password } = validatedFields.data;
-
-          const user = await prismaClient.user.findUnique({
-            where: { email },
-          });
-
-          if (!user) {
-            return null;
-          }
-
-          if (!user.password) {
-            return null;
-          }
-
-          const isValid = await bcrypt.compare(password, user.password);
-
-          if (!isValid) {
-            return null;
-          }
-
-          return user;
-        }
-
-        return null;
-      },
-    }),
   ],
-} ;
+  session: {
+    strategy: 'jwt',
+    maxAge: 60 * 60 * 24, // 24 hours
+  },
+  secret: process.env.NEXTAUTH_SECRET,
+  callbacks: {
+    async session({ session, token }) {
+      if (session.user) {
+        session.user.name = token.name;
+        session.user.isExistingUser = token.isExistingUser;
+        session.user.role = token.role;
+        session.user.isOnboarded = token.isOnboarded;
+        session.user.id = token.id;
+        session.user.eventId = token.eventId;
+      }
 
-export default authOptions;
+      return session;
+    },
+  },
+};
+
+export default authConfig;
